+++ Raycoon.RayMigrator.Database.Sqlite/Templates/Repository_CheckCreate.sql
@@ -254,6 +254,7 @@ INSERT OR IGNORE INTO "{CFG:TableBaseName}MigrationOperation" ("Id", "Name", "De
 
 INSERT OR IGNORE INTO "{CFG:TableBaseName}MigrationRunResult" ("Id", "Name", "Description") VALUES
     (10, 'Running', 'Migration process is currently running'),
+    (50, 'PartialSuccess', 'Migration(s) finished but at least one file was skipped or left Failed'),
     (90, 'Error', 'Migration(s) stopped due to error(s)'),
     (100, 'Ok', 'Migration(s) successfully executed');